the big one
This adds nearly everything that will be done in this version of the prototype.

This project shows off the functionality of my battle system. This version allows adding effects that either last forever, or expire after X turns. Also, once an entity HP reaches 0, it gains the downed status effect.
Downed entities can't be targeted or act.
if all entities on a side are downed then the battle ends.

This also features sprite animations that can either track a line towards the target or play directly on top of the target.